Split import_edge.hpp: node_based_edge.hpp, edge_based_edge.hpp closes #1604
